Move #invoke_callbacks to a concern.
Add callback hooks for Admin::PaymentsController#create.

Allows extending payment controller for new payment methods such as gift vouchers.
