+++ backend/lib/spree/backend.rb
@@ -7,4 +7,6 @@ require 'select2-rails'
 require 'spree/core'
 
 require 'spree/responder'
+require 'spree/backend/action_callbacks'
+require 'spree/backend/callbacks'
 require 'spree/backend/engine'
