+++ backend/lib/spree/backend/action_callbacks.rb
@@ -21,6 +21,6 @@ module Spree
     def fails(method)
       @fails_methods << method
     end
-  end
 
+  end
 end
